requirements.txt
notion-client
openai
PyGithub
python-slugify
requests
pillow
